chore(Storybook): Isolate the icon arg types in their own module
They pull the entire icon package into every chunk that imports the
shared arg types, while most stories only use the lightweight ones.
Stories without an icon control no longer pay that cost.
