+++ storybook/src/components/Badge/Badge.stories.tsx
@@ -9,7 +9,8 @@ import * as Icons from '@amsterdam/design-system-react-icons'
 import { Badge } from '@amsterdam/design-system-react/src'
 import { badgeColors } from '@amsterdam/design-system-react/src/Badge/Badge'
 
-import { colorArgType, iconArgType } from '#storybook/_common/argTypes'
+import { colorArgType } from '#storybook/_common/argTypes'
+import { iconArgType } from '#storybook/_common/iconArgTypes'
 
 const meta = {
   title: 'Components/Feedback/Badge',
