+++ storybook/src/components/Menu/MenuLink.stories.tsx
@@ -8,7 +8,8 @@ import type { Meta, StoryObj } from '@storybook/react-vite'
 import { PieChartFillIcon } from '@amsterdam/design-system-react-icons'
 import { Menu } from '@amsterdam/design-system-react/src'
 
-import { childrenArgType, hrefArgType, linkComponentArgType, requiredIconArgType } from '#storybook/_common/argTypes'
+import { childrenArgType, hrefArgType, linkComponentArgType } from '#storybook/_common/argTypes'
+import { requiredIconArgType } from '#storybook/_common/iconArgTypes'
 
 const meta = {
   title: 'Components/Navigation/Menu',
