+++ storybook/src/components/PageHeader/PageHeaderMenuLink.stories.tsx
@@ -7,7 +7,8 @@ import type { Meta, StoryObj } from '@storybook/react-vite'
 
 import { PageHeader } from '@amsterdam/design-system-react/src'
 
-import { childrenArgType, hrefArgType, iconArgType, linkComponentArgType } from '#storybook/_common/argTypes'
+import { childrenArgType, hrefArgType, linkComponentArgType } from '#storybook/_common/argTypes'
+import { iconArgType } from '#storybook/_common/iconArgTypes'
 
 const meta = {
   title: 'Components/Containers/Page Header',
