+++ storybook/src/components/TabNavigation/TabNavigationLink.stories.tsx
@@ -7,7 +7,8 @@ import type { Meta, StoryObj } from '@storybook/react-vite'
 
 import { TabNavigation } from '@amsterdam/design-system-react/src'
 
-import { childrenArgType, hrefArgType, iconArgType, linkComponentArgType } from '#storybook/_common/argTypes'
+import { childrenArgType, hrefArgType, linkComponentArgType } from '#storybook/_common/argTypes'
+import { iconArgType } from '#storybook/_common/iconArgTypes'
 
 const meta = {
   title: 'Components/Navigation/Tab Navigation',
